requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Message Queue
pika==1.3.2

# gRPC
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1

# HTTP/2 Support
hypercorn==0.15.0
h2==4.1.0

# Video Processing
ffmpeg-python==0.2.0

# Caching
redis==5.0.1

# Testing
locust==2.19.1
pytest==7.4.3

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0